rename components, fix routes
