+++ frontend/src/components/PageNotFound.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const PageNotFound = () => {
+  return (
+    <div>
+      <p>404 - Page couldn't be found</p>
+    </div>
+  );
+};
+
+export { PageNotFound as default };
+
+// The 404 page
