+++ frontend/src/components/UserProfile.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+import Search from "./Search";
+
+const UserProfile = () => {
+  return (
+    <div>
+      <h1>Hello Anna!</h1>
+      <p>Let's find your perfect wine</p>
+      <Search />
+    </div>
+  );
+};
+
+export { UserProfile as default };
+
+// The Profile Page of the User
