--- frontend/src/components/about-us.js
@@ -1,30 +1,0 @@
-import React from 'react';
-
-const AboutUs = () => {
-return (
-    <div>
-        <h1>About Us</h1>
-        <h2>Learn more about the team.</h2>
-        <div>
-        <h3>Data Science</h3>
-        <p>Marc Hendel & Juliana Gomes</p>
-        </div>
-        <div>
-        <h3>UX Design</h3>
-        <p>Ebunoluwa Meroyi, Berenika Godlewska & Ngoc Anh Doan Thi</p>
-        </div>
-        <div>
-        <h3>Web Development</h3>
-        <p>Timea Varju & Franziska Pendzialek</p>
-        </div>
-        <div>
-        <h2>About TechLabs</h2>
-        <p>We dream of a world full of digital shapers. Individuals who use technological tools to approach the challenges of our time with a digital and entrepreneurial mindset.</p>
-        <a href="https://www.techlabs.org/about">More about TechLabs</a>
-        </div>
-    </div>
-)};
-    
-    export {AboutUs as default };
-
-    //This is an about us page.
