--- frontend/src/components/mainpage.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Search from './search'
-
-const MainPage = () => {
-return (
-<div>
-<h1>You're searching for the right wine?</h1>
-<p>Enter your choice here and get a recommendation!</p>
-<Search />
-</div>
-)  
-};
-
-export {MainPage as default };
-
-// The Main Paige of the Website
