--- frontend/src/components/pagenotfound.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const PageNotFound = () => {
-    return (
-    <div>
-    <p>404 - Page couldn't be found</p>
-    </div>
-    )  
-    };
-    
-    export {PageNotFound as default };
-
-
-    // The 404 page
