--- frontend/src/components/results.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Results = () => {
-    return (
-        <div>
-        <h1>Wohoo! Here are your results</h1>
-        <p>Your wine should most like be</p>
-        <ul>
-        <li>a dry</li>
-        <li>red wine</li>
-        <li>from Spain</li>
-        <li>probably a Temperanillo</li>
-        <li>in the price range of X€ to Y€</li>
-        <li>to eat with "food".</li>
-        </ul>
-
-        <p>Here are your wines to check out:</p>
-        
-        </div>
-    )
-}
-
-export {Results as default };
-
-// This component is made to show the results of the search
