--- frontend/src/components/userprofile.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const UserProfile = () => {
-return (
-<div>
-<h1>Your Profile</h1>
-<p>Enter your choice here and get a recommendation!</p>
-</div>
-)  
-};
-
-export {UserProfile as default };
-
-// The Profile Page of the User
